Fix: Converting circular structure to JSON (#12)
