+++ package.json
@@ -101,6 +101,7 @@
     "dependencies": {
         "date-fns": "^2.25.0",
         "geckodriver": "^1.10.0",
+        "json-stringify-safe": "^5.0.1",
         "lodash": "^4.17.21"
     }
 }
