+++ packages/playground/src/todo-6/stores/RootStore.js
@@ -1,0 +1,5 @@
+import TodoStore from './TodoStore';
+
+export default class RootStore {
+  todoStore = new TodoStore(this);
+}
