+++ src/backend/mobxLog.js
@@ -1,5 +1,6 @@
 import { format } from 'date-fns';
 import { cloneDeep } from 'lodash';
+import stringify from 'json-stringify-safe';
 import makeChangesProcessor from '../utils/changesProcessor';
 import consoleLogChange from './utils/consoleLogChange';
 import makeInspector from './utils/inspector';
@@ -10,7 +11,7 @@ const getStoreDataFromChangeObj = obj => {
   let storeData = {};
   if (obj) {
     try {
-      storeData = JSON.parse(JSON.stringify(cloneDeep(obj)));
+      storeData = JSON.parse(stringify(cloneDeep(obj)));
     } catch (err) {
       storeData = { error: err.message };
     }
@@ -33,7 +34,7 @@ const summary = change => {
   sum.removedCount = change.removedCount;
   sum.object = change.object;
   // the newly added
-  sum.storeName = change.object.constructor && change.object.constructor.name;
+  sum.storeName = change && change.object && change.object.constructor && change.object.constructor.name;
   sum.actionName = change.name;
   sum.time = format(new Date(change.timestamp), 'HH:mm:ss');
   sum.storeData = getStoreDataFromChangeObj(change.object);
